+++ actiontext/app/models/action_text/rich_text.rb
@@ -5,7 +5,7 @@
 module ActionText
   # # Action Text RichText
   #
-  # The RichText record holds the content produced by the Trix editor in a
+  # The RichText record holds the content produced by the editor in a
   # serialized `body` attribute. It also holds all the references to the embedded
   # files, which are stored using Active Storage. This record is then associated
   # with the Active Record model the application desires to have rich text content