+++ actiontext/test/dummy/config/importmap.rb
@@ -4,3 +4,5 @@ pin "application", preload: true
 pin "@hotwired/turbo-rails", to: "turbo.min.js", preload: true
 pin "trix"
 pin "@rails/actiontext", to: "actiontext.esm.js"
+pin "lexxy", to: "lexxy.js"
+pin "@rails/activestorage", to: "activestorage.esm.js"